pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "happykube"
version = "2.0.0"
description = "AI-powered emotional intelligence Telegram bot with enterprise architecture"
authors = [{name = "Marco Bellingeri", email = "marco@example.com"}]
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
keywords = ["telegram", "bot", "emotion", "sentiment", "nlp", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "redis>=5.2.0",
    "python-telegram-bot>=21.7",
    "transformers>=4.46.0",
    "torch>=2.5.1",
    "tokenizers>=0.20.3",
    "safetensors>=0.4.5",
    "cryptography>=43.0.3",
    "python-dotenv>=1.0.1",
    "prometheus-client>=0.21.0",
    "flask-cors>=5.0.0",
    "flask-limiter>=3.8.0",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.1",
    "bleach>=6.2.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.28.1",
    "black>=24.10.0",
    "ruff>=0.8.4",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "safety>=3.2.11",
    "pre-commit>=4.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # bandit
]
ignore = [
    "E501",  # line too long (handled by black)
    "S101",  # assert used (ok in tests)
]

[tool.ruff.per-file-ignores]
"tests/**" = ["S101", "S106"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "transformers.*",
    "torch.*",
    "telegram.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
